backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pydantic[email]==2.10.3
pydantic-settings==2.6.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
openai==1.55.3
youtube-transcript-api==0.6.3
python-dotenv==1.0.1
alembic==1.14.0
bcrypt==4.2.0
PyJWT==2.10.1
boto3==1.35.0
